requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10.1
MarkupSafe==1.0
Pillow==7.0.0
Werkzeug==0.15.3

